+++ packages/dsw-document-worker/dsw/document_worker/templates/steps/conversion.py
@@ -27,7 +27,10 @@ class WkHtmlToPdfStep(Step):
             metadata=self.options,
             workdir=str(self.template.template_dir),
         )
-        return DocumentFile(self.OUTPUT_FORMAT, data)
+        return DocumentFile(
+            file_format=self.OUTPUT_FORMAT,
+            content=data,
+        )
 
 
 class WeasyPrintStep(Step):
@@ -74,7 +77,10 @@ class WeasyPrintStep(Step):
             version=self.wp_version,
             custom_metadata=self.wp_custom_metadata,
         )
-        return DocumentFile(self.OUTPUT_FORMAT, data)
+        return DocumentFile(
+            file_format=self.OUTPUT_FORMAT,
+            content=data,
+        )
 
 
 class PandocStep(Step):
@@ -102,6 +108,19 @@ class PandocStep(Step):
         FileFormats.RST,
         FileFormats.RTF,
     ])
+    OUTPUT_ENCODINGS = {
+        FileFormats.ADoc: DEFAULT_ENCODING,
+        FileFormats.DocBook4: None,
+        FileFormats.DocBook5: None,
+        FileFormats.DOCX: None,
+        FileFormats.EPUB: None,
+        FileFormats.HTML: DEFAULT_ENCODING,
+        FileFormats.LaTeX: DEFAULT_ENCODING,
+        FileFormats.Markdown: DEFAULT_ENCODING,
+        FileFormats.ODT: None,
+        FileFormats.RST: DEFAULT_ENCODING,
+        FileFormats.RTF: None,
+    }
 
     OPTION_FROM = 'from'
     OPTION_TO = 'to'
@@ -129,7 +148,11 @@ class PandocStep(Step):
             metadata=self.options,
             workdir=str(self.template.template_dir),
         )
-        return DocumentFile(self.output_format, data)
+        return DocumentFile(
+            file_format=self.output_format,
+            content=data,
+            encoding=self.OUTPUT_ENCODINGS[self.output_format],
+        )
 
 
 class RdfLibConvertStep(Step):
@@ -170,7 +193,11 @@ class RdfLibConvertStep(Step):
         data = self.rdflib_convert(
             self.input_format, self.output_format, document.content, self.options
         )
-        return DocumentFile(self.output_format, data)
+        return DocumentFile(
+            file_format=self.output_format,
+            content=data,
+            encoding=DEFAULT_ENCODING,
+        )
 
 
 register_step(WkHtmlToPdfStep.NAME, WkHtmlToPdfStep)